Added banner, altered layout, changed index
