+++ components/card/cardContainer.js
@@ -2,7 +2,6 @@ import Card from "./card";
 import styles from "./cardContainer.module.css";
 
 export default function CardContainer({ recipeCards }) {
-  //console.log(recipeCards);
   return (
     <div className={styles.card}>
       {recipeCards.map(({ id, name, description, image }) => (
