+++ components/layout.module.css
@@ -1,5 +1,5 @@
 .container {
-  max-width: 36rem;
+  max-width: 75rem;
   padding: 0 1rem;
   margin: 3rem auto 6rem;
 }
