+++ styles/utils.module.css
@@ -51,3 +51,11 @@
 .lightText {
   color: #666;
 }
+
+.headingCentered {
+  font-size: 1.5rem;
+  line-height: 1.4;
+  margin: 1rem 0;
+  text-align: center;
+  padding: 0.5rem;
+}
